+++ apps/mobile/app.config.ts
@@ -1,6 +1,6 @@
 import { ConfigContext, ExpoConfig } from 'expo/config';
 
-const version = '2.113.0'; // x-release-please-version
+const version = '2.114.0'; // x-release-please-version
 
 export default ({ config }: ConfigContext): ExpoConfig => {
   return {