+++ packages/queries/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.15.4](https://github.com/leather-io/mono/compare/@leather.io/queries-v0.15.3...@leather.io/queries-v0.15.4) (2026-08-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/models bumped to 0.61.1
+    * @leather.io/services bumped to 1.64.1
+    * @leather.io/utils bumped to 0.52.7
+
 ## [0.15.3](https://github.com/leather-io/mono/compare/@leather.io/queries-v0.15.2...@leather.io/queries-v0.15.3) (2026-08-18)